build fix

+++ src/10_db/102_mysql/MySQLDB.cpp
@@ -871,7 +871,7 @@ namespace synthese
 		void MySQLDB::DBRecordCellBindConvertor::operator()( const size_t& s ) const
 		{
 			_bnd.buffer_type = MYSQL_TYPE_LONG;
-			_bnd.buffer = static_cast<void*>(const_cast<int*>(static_cast<const size_t*>(&s)));
+			_bnd.buffer = static_cast<void*>(const_cast<int*>(static_cast<const int*>(&s)));
 			_bnd.buffer_length = 0;
 			_bnd.is_null_value = false;
 		}
